% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{af}
\alias{af}
\title{af}
\usage{
af(altc, totc, grid.size = 0.01, verbose = F)
}
\description{
computes prob density of af over a set of n mutation calls
using alt allele count, total counts, cn, and provided
grid.size

as per landau, carter et al

altc, totc, and cn are of length n, purity is length 1
}
\author{
Marcin Imielinski
}

