% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{af2}
\alias{af2}
\title{af2}
\usage{
af2(altc1, totc1, altc2, totc2, grid.size = 0.01, verbose = F,
  animate = NA)
}
\description{
computes 2D probability density af over a set of n mutation counts
two samples, output columns correspond to x and rows correspond to y
and dimnames correspond to amounts

as per landau, carter et al

altc1, totc1, altc2, totc2 are of length n
}
\author{
Marcin Imielinski
}
