% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{affine.map}
\alias{affine.map}
\title{affine.map}
\usage{
affine.map(x, ylim = c(0, 1), xlim = c(min(x), max(x)), cap = F,
  cap.min = cap, cap.max = cap, clip = T, clip.min = clip,
  clip.max = clip)
}
\arguments{
\item{x}{input vector to affinely transform}

\item{ylim}{length 2 vector of y bounds to project data into (=c(0,1))}

\item{xlim}{length 2 vector of x bounds to project data from (=c(min(x), max(x)))}

\item{cap}{logical flag whether to cap data below xmin and above xmax of x vector ie return as ymin, ymax respectively (=FALSE)}

\item{cap.min}{logical flag whether to cap data below xmin  of x vector ie return as ymin (=cap)}

\item{cap.max}{logical flag whether to cap  data above xmax of x vector ie return as ymax (=cap)}

\item{clip}{logical flag whether to clip data below xmin and above xmax of x vector ie return as NA (=TRUE)}

\item{clip.min}{logical flag whether to clip data below xmin  of x vector ie return as NA (=clip)}

\item{clip.max}{logical flag whether to clip data above xmax  of x vector ie return as NA (=clip)}
}
\value{
x values transformed by affine map defined by xlim and ylim
}
\description{
affinely maps 1D points in vector x from interval xlim to interval ylim,
ie takes points that lie in 
interval xlim and mapping onto interval ylim using linear / affine map defined by:
(x0,y0) = c(xlim(1), ylim(1)),
(x1,y1) = c(xlim(2), ylim(2))
(using two point formula for line)
useful for plotting.

if cap.max or cap.min == T then values outside of the range will be capped at min or max
}
\author{
Marcin Imielinski
}

