% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{alignment_metrics}
\alias{alignment_metrics}
\title{alignment_metrics}
\usage{
alignment_metrics(job)
}
\description{
Grabs alignment metrics from BWAMemFast like Job object (e.g. BWA mem fast)
or paths to bam files aligned with some variant of alignment pipelines run with
Picard alignment_summary_metrics postprocessing. 
and returns the category = 'PAIR' rows of alignment_summary_metrics indexed by entity id
}
\author{
Marcin Imielinski
}
