% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{allele_multiplicity}
\alias{allele_multiplicity}
\title{allele_multiplicity}
\usage{
allele_multiplicity(ind, maf, abs.seg = NULL, verbose = TRUE)
}
\description{
Given individuals file (with either Purity value or a bsp_participant_id or Tumor_scan_name columns)
and maf file annotated with absolut copy number (maf$cn.tot) and column $patient for firehose id.

You can provide maf without cn.tot annotations latter case will pull absolute segs
using bsp_participant_id in ind and annotate, but may take a few minutes to pull.
}
\author{
Marcin Imielinski
}

