% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{anno.hops}
\alias{anno.hops}
\alias{anno.hop}
\title{anno.hops}
\usage{
anno.hop(walks)
}
\arguments{
\item{walks}{walks to annotate}
}
\description{
Adds simple annotations to GRangesList of walks including
distance along each reference fragment and distance
between "hops"
}
\author{
Marcin Imielinski
}
