% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{arrstring}
\alias{arrstring}
\title{arrstring}
\usage{
arrstring(A, x = NULL, sep = ", ", sep2 = "_", signif = 3, dt = FALSE)
}
\arguments{
\item{A}{array}

\item{sep}{separator to use between table elements}
}
\value{
character representation of table
}
\description{
string representation of row array as linear combination of nonzero entries
of that row using column names as variables
}
\author{
Marcin Imielinski
}
