% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{bisort}
\alias{bisort}
\title{bisort}
\usage{
bisort(A, drop = F)
}
\arguments{
\item{A}{matrix to sort}

\item{drop}{logical flag whether to drop empty rows (=FALSE)}
}
\description{
"bisorts" matrix according to rows and columns (and optionally removes empty rows, ie with no nonzero)
}
\author{
Marcin Imielinski
}

