% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{border}
\alias{border}
\title{border}
\usage{
border(B, na.rm = TRUE)
}
\arguments{
\item{B}{input matrix logical format, or castable to logical}

\item{na.rm}{removes NA}
}
\value{
B resorted using "binary" order
}
\description{
orders rows of a logical / binary matrix treating each row as binary number with digits encoded as TRUE / FALSE values of entries
}
