% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{bplot}
\alias{bplot}
\title{bplot}
\usage{
bplot(
  y,
  by = NULL,
  facet = NULL,
  col = NA,
  ylim = NA,
  keep.base = NULL,
  xlab = "",
  counts = TRUE,
  wes = TRUE,
  ylab = "",
  title = "",
  print = TRUE,
  return.stats = FALSE
)
}
\arguments{
\item{y}{categorical vector across N samples whose fraction will be plotted on the y axis}

\item{by}{optional grouping vector, same length as y, to group on prior to plotting fractions}

\item{facet}{optional second grouping vector, same length as y, to group on}

\item{col}{single color or named vector to plot or NA (then brewer is used)}

\item{keep.base}{whether to keep the base level when plotting, will be FALSE unless y is logical (ie suppress plotting FALSE value)}

\item{xlab}{x label}

\item{counts}{print counts alongside the by / facet labels (TRUE)}

\item{wes}{if col = NA wes = TRUE will cause wesanderson colors to be used, otherwise standard brewer is used (TRUE)}

\item{ylab}{y label}

\item{title}{title}

\item{return.stats}{return stats with confidence intervals and proportions in each tranche}
}
\value{
gg plot object or stats
}
\description{
Takes vector y of binary / categorical data (eg genotype) plots a bar plot of the
fraction +/- confidence intervals (from prop.test) of each unique value in y. 

Optional argument by and facet provide additonal grouping / faceting variables that will
result in the fractions being computed within those groups and the results plotted vertically
(for by = ) and horizontally (for facet = ).
}
\author{
Marcin Imielinski
}
