% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{bsub_cmd}
\alias{bsub_cmd}
\title{bsub_cmd}
\usage{
bsub_cmd(cmd, queue, jname = NULL, jlabel = NULL, jgroup = NULL,
  mem = NULL, group = "cgafolk", cwd = NULL, mc.cores = NULL,
  deadline = F)
}
\arguments{
\item{cmd}{length n  vector of shell commands, optionally named, one per job}

\item{queue}{optional length n or length 1 character specifying queue to send jobs to (default hour)}

\item{jname}{optional length n character specifying names of jobs, this will be the root of the output files generated by the job}

\item{jlabel}{optional length n character specifying labels of jobs, this the string}

\item{jgroup}{optional length n character specifying job group name}

\item{mem}{length n or length 1 integer specifying GB of memory to be used by jobs}

\item{group}{character specifying job group (default cgafolk)}

\item{cwd}{character specifying which working directory to launch jobs from (default is current working directory of R session)}

\item{mc.cores}{length n or 1 integer specifying how many cores to assign to each job}

\item{deadline}{logical flag whether to send jobs to deadline queue}
}
\value{
character vector of bsub commands, which can run using system or dumped to a shell script
}
\description{
Makes bsub command that wraps shell command "cmd" to send to queue "queue"
redirebmccting output / error etc streams to path prefixed by "jname",
optional_args: maximum memory requirements "mem", "jlabel" job label
b
}
\author{
Marcin Imielinski
}
