% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{bubblemap}
\alias{bubblemap}
\title{bubblemap}
\usage{
bubblemap(
  mat,
  col = "darkgreen",
  cluster = TRUE,
  cex = 1,
  cex.text = 1,
  zlim = cex * c(0, 10),
  col.text = "white",
  show.legend = FALSE
)
}
\description{
Quick bubble heatmap from matrix mat, use cex to tweak bubble size.
}
