% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{camerplot}
\alias{camerplot}
\alias{cameraplot}
\title{cameraplot}
\usage{
cameraplot(camera.res, gene.sets, voom.res, design, contrast = ncol(design),
  title = "Camera Gene set notch plot", cex.space = 1,
  col.axis = alpha("gray20", 0.8), col.ramp = c("blue", "red"),
  cex.slab = 1, cex.glab = 1, lwd.notch = 1, tick.w = 0.1,
  max.genes = 10, text.shift = 0.5, height.wf = 0.1, min.corr = 0.1,
  min.dist = 10, max.gene.sets = 20, gtext.shift = 0.2)
}
\arguments{
\item{camera.res}{output of camera from limma}

\item{gene.sets}{gene set input to camera (named list of indices into the voom.res gene expression matrix)}

\item{voom.res}{output of voom from limma}

\item{design}{design matrix input to camera}

\item{cex.space}{label spacing expansion factor (use if labels get too crowded}

\item{col.axis}{axis color character}

\item{col.ramp}{ramp from lowest to highest expression to phenotype correlation (default blue, red)}

\item{cex.slab}{set label cex}

\item{cex.glab}{gene label cex}

\item{lwd.notch}{notch thickness}

\item{max.genes}{max genes to draw in "leading edge" of gene set}

\item{text.shift}{amount to shift text from notches (>0, <1)}

\item{height.wf}{height of the topmost correlation waterfall plot}

\item{min.corr}{minimal abs(correlation) value for leading edge definition}

\item{text.shift}{minimal distance between labels}
}
\description{
plots the results of CAMERA in limma package
}
\author{
Marcin Imielinski
}
