% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{cc}
\alias{cc}
\title{cc}
\usage{
cc(x, y = c(), ...)
}
\arguments{
\item{x}{column expression or variables}

\item{...}{additional variables}
}
\description{
select columns of data.frame or data.frame-like object
}
