% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{ccf}
\alias{ccf}
\title{ccf}
\usage{
ccf(altc, totc, cn, purity, grid.size = 0.01, verbose = F)
}
\description{
computes fuzzy histogram of ccf across a set of n mutation calls
using alt allele count, total counts, cn, and purity, and provided
grid.size

as per landau, carter et al

altc, totc, and cn are of length n, purity is length
}
\author{
Marcin Imielinski
}

