% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{charToDec}
\alias{charToDec}
\title{charToDec}
\usage{
charToDec(c)
}
\arguments{
\item{c}{character vector}
}
\value{
length(c) integer vector of byte representation of c
}
\description{
converts character vector to byte vector in decimal representation
}
\author{
Marcin Imielinski
}
