% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{chunk}
\alias{chunk}
\title{chunk}
\usage{
chunk(from, to = NULL, by = 1, length.out = NULL)
}
\arguments{
\item{from}{integer where to begin sequence}

\item{to}{integer to end sequence}

\item{by}{interval to space sequence}

\item{length.out}{number of desired chunks, i.e. nrows of output matrix}
}
\value{
2 column matrix of indices, each row representing chunk
}
\description{
takes same input as seq (from, to, by, length.out) and outputs a 2 column matrix of indices
corresponding to "chunks"
}
\author{
Marcin Imielinski
}
