% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{clock}
\alias{clock}
\title{clock

times expression}
\usage{
clock(expr)
}
\arguments{
\item{expr}{R code to eval while suppressing all errors}
}
\description{
clock

times expression
}
\author{
Marcin Imielinski
}
