% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{clone_cluster}
\alias{clone_cluster}
\title{clone_cluster}
\usage{
clone_cluster(
  altc,
  totc,
  cn,
  purity,
  thresh = 0.95,
  k = 2,
  mix.model = NULL,
  grid.size = 0.01,
  verbose = F,
  nsamp = 10000
)
}
\description{
determines "clone membership" using CCF threshold for n variants, given alt read count, total read count, and purity

first fits k component mixture model (k pre-specified) to CCF histogram .. mixture model can also be given as input
returns cluster centers and membership probabilities for each mutation

$mu k vector of means
$sigma k vector of sigma
$p.cluster n x k matrix of cluster probabilities
$lambda k vector of cluster membership frequencies
return n x k matrix of probabilities that CCF>ccf.thresh for each variant

altc, totc, and cn are of length n, purity is length 1
}
\author{
Marcin Imielinski
}
