% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{col.slice}
\alias{\%!\%}
\alias{col.slice}
\title{col.slice}
\usage{
df \%!\% ...
}
\arguments{
\item{df}{data.frame}

\item{regex}{string to match or number in which case that column is returned (same behavior for data.table)}
}
\value{
slices of data.frame matching regex
}
\description{
Hacked operator for subsetting columns of data.frames, DataFrames, data.tables, GRanges using
a vector of regexps

df %!% c('string.*to.*match', 'another.string.to.match')
}
\author{
Marcin Imielinski
}

