% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\docType{class}
\name{coloredData}
\alias{coloredData}
\title{coloredData}
\usage{
coloredData(data, colormap, upright = T)
}
\description{
S4 class for data with colors used by heatmap.plus

simple object with data (e.g. vector or matrix of categorical, real numbers) + a colormap

 colormap is a (named) vector mapping factor levels / unique values in data into colors,
 or otherwise assigning a color range to numeric data.

Instantiate coloredData
}
\author{
Marcin Imielinski
}
