% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{complete}
\alias{complete}
\title{complete}
\usage{
complete(d, rows = FALSE)
}
\description{
outputs columns that are "complete" for all rows
or (if rows = TRUE) rows that are complete for columns
}
\author{
Marcin Imielinski
}
