% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{cyto2igraph}
\alias{cyto2igraph}
\title{cyto2igraph}
\usage{
cyto2igraph(cw)
}
\arguments{
\item{cw}{CytoscapeWindow object to grab from (see Rcytoscape)}
}
\value{
igraph object
}
\description{
Pulls graph in CytoscapeWindow "cw" and returns as igraph object
}
\author{
Marcin Imielinski
}

