% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{cytoscape}
\alias{cytoscape}
\title{cytoscape}
\usage{
cytoscape(graph = NULL, sessionName = "M-ski",
  host = Sys.getenv("LOCAL.COMPUTER"), port = 9000, display = T,
  layout = "degree-circle", verbose = T, ...)
}
\arguments{
\item{graph}{igraph or adjacency matrix}

\item{sessionName}{character name of cytoscape session to open on local computer}

\item{host}{character name of host on which cytoscape is running (=Sys.getenv('LOCAL.COMPUTER'))}

\item{port}{integer port on which cytoscape is running (=9000)}

\item{display}{logical flag whether to display graph locally (=TRUE)}

\item{layout}{character specifying layout to display as (=degree-circle)}

\item{verbose}{logical vector whether to use verbose output}

\item{...}{additional arguments to new.CytoscapeWindow}
}
\description{
shortcut to connect to local cytoscape instance running on LOCAL.COMPUTER (unix environment variable) via RPC call

graph must be igraph, or adjacency matrixx
}
\author{
Marcin Imielinski
}

