% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{d3igraph}
\alias{d3igraph}
\title{d3igraph}
\usage{
d3igraph(g)
}
\description{
Wrapper around network d3 package to quickly convert igraph to D3 and visualize
Note: note can send output to plot.html via vij
}
