% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{dcast.data.table2}
\alias{dcast.data.table2}
\title{dcast.data.table but allows vector arguments for value.var,}
\usage{
dcast.data.table2(data, formula, ..., value.var = NULL, sep = "_")
}
\description{
if value.var is a vector then will combine the right hand side column names with each element of value.var
in a merged cast table
}
\author{
Marcin Imielinski
}

