% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{dcast2}
\alias{dcast2}
\title{dcast.data.table but allows vector arguments for value.var,}
\usage{
dcast2(data, formula, ..., value.var = NULL, fun.aggregate = function(x) if
  (length(x) <= 1) x[1] else paste(x, collapse = ","), sep = "_")
}
\description{
if value.var is a vector then will combine the right hand side column names with each element of value.var
in a merged cast table
}
\author{
Marcin Imielinski
}

