% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{dfstring}
\alias{dfstring}
\title{dfstring}
\usage{
dfstring(df, oneline = TRUE, sep1 = "; ", sep2 = ", ")
}
\arguments{
\item{df}{data.frame input}

\item{oneline}{logical flag whether to print on one line (=TRUE)}

\item{sep1}{first level separator (=;) i.e. between rows}

\item{sep2}{second level separator (=, ) i.e. between columns}
}
\value{
character vector of string representation
}
\description{
"tuple" style chraacter representation of a table, key name1 = value1, name2 = value2
either as a single line or many lines
useful for quick eyeballing of tabular data
}
\author{
Marcin Imielinski
}

