% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{dirr}
\alias{dirr}
\title{dirr}
\usage{
dirr(x, pattern = NULL, rep = "", full = TRUE, ...)
}
\arguments{
\item{x}{character of path to run dir on}

\item{pattern}{character pattern to limit files to and to replace with rep}

\item{rep}{character pattern to replace filenames with}

\item{full}{whether to return full path}

\item{...}{additional arguments to dir}
}
\value{
named vector of file paths, named by file names in dir gsub-stripped with pattern
}
\description{
a variant of dir that gsubs pattern from normal output of dir to name output vector

eg dirr(path, '.txt') will return dir output with .txt removed
eg dirr(path, '.txt', '.rds' ) will return dir output with .txt subbed with .rds
}
\author{
Marcin Imielinski
}

