% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{dplot}
\alias{dplot}
\title{dplot}
\usage{
dplot(y, group, ylab = "", xlab = "", log = F, dotsize = NULL,
  binwidth = NULL, title = NULL, ylim = NULL, text.size = NULL)
}
\arguments{
\item{y}{numeric vector of data}

\item{group}{length(y) vector of catageories}

\item{ylab}{y axis label (='')}

\item{xlab}{x axis label (='')}

\item{log}{logical flag whether to plot y axis in log format (=FALSE)}

\item{dotsize}{integer dot size to plot with (= NULL)}

\item{binwidth}{numeric binwidth of histogram (= NULL)}

\item{title}{character title of plot (='')}

\item{ylim}{y limits of plot (= NULL)}

\item{text.size}{text size of legend (= NULL)}
}
\description{
Plots dotplot of grouped data
}
\author{
Marcin Imielinski
}

