% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{dtt}
\alias{dtt}
\title{dtt}
\usage{
dtt(threads = 1)
}
\arguments{
\item{numthreads}{number of threads to set setDTthreads}
}
\description{
Easy to type shortcut for setDTthreads
}
\author{
Marcin Imielinski
}
