% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{dump_dranger_for_IGV_dump}
\alias{dump_dranger_for_IGV_dump}
\alias{dump_dranger_for_IGV_snapshots}
\title{_dranger_for_IGV_snapshots}
\usage{
dump_dranger_for_IGV_snapshots(dranger, iset, wkspace = "Lung",
  igv.dir = "./", mutation.log.name = "mutation_log",
  snapshots.queue.filename = "snapshot_queue", dump.mutation.log = TRUE,
  chr.include = T)
}
\description{
dumps "snapshot_queue" file from dranger table into igv.dir for IGV screen shots of translocations
Takes as input:
dranger = data frame obtained by reading in dranger output
iset= firehose iset name OR data.frame of individuals table from firehose obtained by doing get_fh(...,  merged = TRUE) which contains columns $Individual_Id,
$Tumor_clean_bam_file_wgs, $Normal_clean_Bam_file_wgs;
wkspace = firehose workspace (needs to be provided if iset is a name of a firehose iset) 

(should be followed by command line call to "make snapshots" or "make snapshots_batch" via Makefile in snapshots directory to take snapshots on the farm
of queued mutations on the farm)
}
\author{
Marcin Imielinski
}

