% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{dump_mutations_for_IGV_snapshots}
\alias{dump_mutations_for_IGV_snapshots}
\title{dump_mutations_for_IGV_snapshots}
\usage{
dump_mutations_for_IGV_snapshots(maf, igv.dir, iset = "LUAD_all_exomes",
  wkspace = "Lung", mutation.log.name = "mutation_log",
  snapshots.queue.filename = "snapshot_queue", dump.mutation.log = TRUE,
  capture = TRUE)
}
\description{
dumps "snapshot_queue" file from maf into igv.dir for IGV screen shots
(should be followed by command line call to "make snapshots" or "make snapshots_batch" via Makefile in snapshots directory to take snapshots on the farm
of queued mutations on the farm)

Optionally dumps "mutation_log" which is a text file input to Mutation Station (Eran Hodis 2011) .. this tdf provides
bsolute paths to pngs (resulting from the predicted outputs of the above) and columns for mutation review.

Should set dump.mutation.log to TRUE for first run and then for subsequent re-runs (ie to re-do LSF failures) should run with
dump.mutation.log = FALSE, so as to not rewrite the mutation log.
}
\author{
Marcin Imielinski
}

