% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{dunlist}
\alias{dunlist}
\title{dunlist}
\usage{
dunlist(x, iid = FALSE)
}
\arguments{
\item{x}{list of vectors, matrices, or data frames}
}
\value{
data.frame of concatenated input data with additional fields $ix and $iix specifying the list item and within-list index from which the given row originated from
}
\description{
unlists a list of vectors, matrices, data.tables into a data.table indexed by the list id
$listid

does fill = TRUE in case the data.tables inside the list do not have compatible column names
}
\author{
Marcin Imielinski9
}
