% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{edge2tip}
\alias{edge2tip}
\title{edge2tip

Returns matrix or data.table mapping edge to tips from ape tree}
\usage{
edge2tip(tree, matrix = TRUE)
}
\description{
edge2tip

Returns matrix or data.table mapping edge to tips from ape tree
}
