% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{elcycles}
\alias{elcycles}
\title{elcycles}
\usage{
elcycles(A)
}
\arguments{
\item{A}{adjacency matrix}
}
\value{
list with fields:
$cycles = list of vertices in elementary cycles
$cycles.eix = list of edges in elementary cycles, where edges are numbered according to the 1D index of adj matrix A
}
\description{
enumerates all elementary cycles in a graph via igraph library

A is either an adjacency matrix or igraph object
}
\author{
Marcin Imielinski
}
