% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{file.dir}
\alias{file.dir}
\title{file.dir}
\usage{
file.dir(paths)
}
\arguments{
\item{paths}{character vector of full paths}
}
\value{
character vector of just file.names
}
\description{
grabs file.dirs from list of paths
}
\author{
Marcin Imielinski
}

