% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{file.name}
\alias{file.name}
\title{file.name}
\usage{
file.name(paths)
}
\arguments{
\item{paths}{character vector of full paths}
}
\value{
character vector of just file names
}
\description{
parses filenames from character vector of paths
}
\author{
Marcin Imielinski
}

