% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{fill.blanks}
\alias{fill.blanks}
\title{fill.blanks}
\usage{
fill.blanks(x)
}
\description{
Takes vector with NAs and "fill in blank" positions i with the value of the last non NA position
}
\author{
Marcin IMielinski
}
