% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{fisher.combined}
\alias{fisher.combined}
\title{fisher.combined}
\usage{
fisher.combined(Ps)
}
\arguments{
\item{Ps}{n x k matrix of p values from k different independent tests}
}
\value{
length n numeric vector of p values
}
\description{
Computes fisher combined p value for a matrix of p values where the columns correspond to individual (independent) tests
rows correspond to hypotheses.
}
\author{
Marcin Imielinski
}
