% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{fisher.pairwise}
\alias{fisher.pairwise}
\title{fisher.pairwise}
\usage{
fisher.pairwise(x, y = x)
}
\arguments{
\item{x}{n x k1 data frame of categorical data on k1 variables}

\item{y}{n x k2 optional data frame of categorical data on k2 variables (= x)}
}
\value{
list with field $p and $or correspodning to k1 x k2 matrices of p values and odds ratios for each pair of tests
}
\description{
Performs fisher test on cols of matrix / df x vs cols of matrix / df y

returns list with ncol(x) by ncol(y) matrices $p and $or denoting the p value and odds ratio of the result of the
fisher test on col i of x and col j of y

If y is not provided, will correlate rows of x with themselves.
}
\author{
Marcin Imielinski
}

