% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{fisher.plot}
\alias{fisher.plot}
\title{Plots fisher contingency table with p value}
\usage{
fisher.plot(O)
}
\arguments{
\item{O}{observed matrix of counts}
}
\description{
Plots fisher contingency table with p value
}

