% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{flow_slurm_status}
\alias{flow_slurm_status}
\title{flow_slurm_status}
\usage{
flow_slurm_status(
  jb,
  return_id = FALSE,
  fields = c("jobid", "jobname", "state", "maxrss", "reqm", "elapsed"),
  return.str = FALSE
)
}
\arguments{
\item{jb}{Flow Job object}

\item{return_id}{return a list with the slurm job IDs for the Flow Job jobs}

\item{fields}{which fields to print (see SLURM docs for details)}

\item{return.str}{use system2 to return the stdout as a list of charachters}
}
\value{
vector
}
\description{
wrapper for sacct (SLURM command line function)
}
\author{
Alon Shaiber
}
