% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{fready}
\alias{fready}
\title{fread with name cleaning}
\usage{
fready(..., pattern = "\\\\W+", sub = "_")
}
\arguments{
\item{pattern}{character (default \\W)}

\item{sub}{character to sub in names (default _)}
}
\value{
data.datalbe
}
\description{
calls fread while cleaning names using provided or default pattern and sub
}

