% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{fuckr}
\alias{fuckr}
\title{fuckr}
\usage{
fuckr()
}
\description{
... what you feel when R is getting on your nerves. Toggles options(error = ) to enable / disable debugging mode.

toggles options error recover / NULL
}
\author{
Marcin Imielinski
}
