% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{func_code}
\alias{func_code}
\title{func_code}
\usage{
func_code(func, sources = c(), ...)
}
\arguments{
\item{func}{function to call}

\item{sources}{files to source}

\item{...}{additional arguments to function which should contain numerical or character vectors or lists of such vectors}
}
\value{
character data of source file containing call to function with arguments hard coded
}
\description{
Produces (simple) R code calling function named "func" with args in list "argv", prepending with
source() call to directories in the vector "sources" if specified.

NOTE: args in ... can be lists or vectors consisting of numerical values or characters.  Lists can have named fields.  
These will be assigned in a "hard coded" way in the Rcode, so these should be ideally scalars or
pretty short vectors / lists. 

For code to run properly, the names of "argv" must correspond to argument names of "func", or
if the list has unnamed fields then they must be ordered in the order of the function args.  

Useful for dumping tmp code files for farming when there are many arguments being passed
}
\author{
Marcin Imielinski
}

