% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{gatk_callvariants}
\alias{gatk_callvariants}
\title{gatk_callvariants}
\usage{
gatk_callvariants(bams, outdir = "./", run = gsub("[^\\\\w+]", "_",
  as.character(Sys.time()), perl = T), queue = "hour", memlimit = 2,
  sep = "/", skipclean = T, chunk = 250000, runtype = "wg",
  dry = FALSE, dcov = 200, bsub = FALSE, vqsr.memlimit = memlimit)
}
\arguments{
\item{bams}{bams to run variant calling on}

\item{outdir}{out directory to output to}

\item{run}{name to give run (=timestamp)}

\item{queue}{LSF queue to run on (='week')}

\item{memlimit}{memory limit in GB}

\item{skipclean}{logical flag whether to skipclean (=TRUE)}

\item{chunk}{integer chunk of variants to run each section on (=250e3)}

\item{runtype}{character run type (='wg')}

\item{dry}{logical flag whether to do dry run (=FALSE)}

\item{dcov}{dcov parameter to GATK (=200)}

\item{bsub}{whether to run on LSF vs local (=FALSE)}
}
\description{
Outputs variant calling pipeline on a set of whole genome (or whole exome) bams to shell script and prints
instructions on how to execute from command line.
}
\author{
Marcin Imielinski
}
