% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{get.pileup.gtrack}
\alias{get.pileup.gtrack}
\title{get.pileup.gtrack}
\usage{
get.pileup.gtrack(sites)
}
\arguments{
\item{sites}{output of het pileup (either a data.table object or a path to the tab-delimited file containing the output)}
}
\value{
gTrack object with two tracks corresponding to normal and tumor minor and major allele frequencies
}
\description{
Get a gTrack object with fraction reads matching the major and minor allele in het pileup counts.
}
\details{
Returns a gTrack object with two tracks one for tumor reads and one for the matched normal.
Minor and major allele are colored by green and red respectively.
}
\author{
Alon Shaiber
}
