% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{get.var.col}
\alias{get.var.col}
\alias{get.varcol}
\title{Simple function storing default variant color scheme}
\usage{
get.varcol()
}
\value{
vector of default variant colors
}
\description{
Simple function storing default variant color scheme

Originally in 'mskilab/bamUtils'
}
