% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{get_dryclean_pon_path}
\alias{get_dryclean_pon_path}
\title{get_dryclean_pon_path}
\usage{
get_dryclean_pon_path(
  pon_name,
  pon_dict =
    "/gpfs/commons/groups/imielinski_lab/DB/modules/dryclean/dryclean_pon_dictionary.tsv"
)
}
\arguments{
\item{pon_name}{name of the PON.}

\item{pon_dict_path}{path to the TAB delimited file.}
}
\value{
detergent path to the PON RDS file.
}
\description{
return the path to the PON RDS (detergent) file corresponding to the provided pon_name
}
\author{
Alon Shaiber
}
