% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{get_dryclean_pon_template}
\alias{get_dryclean_pon_template}
\title{get_dryclean_pon_template}
\usage{
get_dryclean_pon_template(
  pon_name,
  pon_dict =
    "/gpfs/commons/groups/imielinski_lab/DB/modules/dryclean/dryclean_pon_dictionary.tsv"
)
}
\arguments{
\item{pon_name}{name of the PON.}

\item{pon_dict_path}{path to the TAB delimited file.}
}
\value{
template GRanges object with the ranges that are compatible with the provided pon_name
}
\description{
return the template GRanges for the PON corresponding to the provided pon_name
}
\author{
Alon Shaiber
}
