% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{get_gene_ampdels_from_jabba}
\alias{get_gene_ampdels_from_jabba}
\title{get_gene_ampdels_from_jabba}
\usage{
get_gene_ampdels_from_jabba(
  jab,
  pge,
  amp.thresh = 4,
  del.thresh = 0.5,
  nseg = NULL
)
}
\arguments{
\item{jab}{either path to jabba_rds output or an object containing the jabba output}

\item{pge}{GRanges of genes (must contain field "gene_name")}

\item{amp.thresh}{threshold to determine amplifications (this is relative to the ploidy normalized copy number. Values are also normalized using the normal copy number so that amplifications and deletions in chromosome X in male samples can be called correctly.}

\item{del.thresh}{threshold to determine deletions.}

\item{nseg}{GRanges with field "ncn" - the normal copy number (if not provided then ncn = 2 is used)}
}
\value{
scna data.table with genes that have either amplification or deletion
}
\description{
Takes a jabba_rds output and returns a GRanges with the genes that have either amplifications or deletions
}
\author{
Alon Shaiber
}
