% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{gigs}
\alias{gigs}
\title{gigs}
\usage{
gigs(x)
}
\description{
Takes string representing memory and returns numeric string representing number of GB
represented by this string or NA
}
\author{
Marcin Imielinski
}
