% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{gr.eval}
\alias{gr.eval}
\title{gr.eval}
\usage{
gr.eval(query, subject, expr, fill = NA, ignore.strand = TRUE)
}
\arguments{
\item{query}{GRanges of intervals that we are interested in populating}

\item{subject}{GRanges of interval that has metadata that we want to aggregate / query}

\item{expr}{expression on subject metadata that we want to evaluate}

\item{fill}{flll value for empty / non overlapping intervas}
}
\value{
vector of length query with expr evaluated on each overlapping interval
}
\description{
Evaluate a scalar expression on the metadata of a subject, returning
a vector of the same length as query with the values populated for
matching value (or fill value for non matching)
}
