% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{gr.match}
\alias{gr.match}
\title{More robust and faster implementation of GenomicRangs::setdiff}
\arguments{
\item{query}{\code{GRanges} object as query}

\item{subject}{\code{GRanges} object as subject}

\item{max.slice}{Default Inf. If query is bigger than this, chunk into smaller on different cores}

\item{verbose}{Default FALSE}

\item{mc.cores}{Default 1. Only works if exceeded max.slice}

\item{...}{arguments to be passed to \link{gr.findoverlaps}}
}
\value{
returns indices of query in subject or NA if none found
}
\description{
Robust to common edge cases of setdiff(gr1, gr2)  where gr2 ranges are contained inside gr1's (yieldings
setdiffs yield two output ranges for some of the input gr1 intervals.
}

