% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{grok_bcf}
\alias{grok_bcf}
\title{Reads and parses bcf via bcftools call}
\usage{
grok_bcf(
  bcf,
  gr = NULL,
  bpath = "/nfs/sw/bcftools/bcftools-1.1/bcftools",
  label = NA,
  filter = "PASS",
  snv = FALSE,
  indel = FALSE,
  het = FALSE,
  hom = FALSE,
  keep.modifier = TRUE,
  long = FALSE,
  oneliner = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{bcf}{path to bcf file}

\item{gr}{optional granges to query}

\item{bpath}{path to bcftools binary executable}
}
\description{
Reads and parses bcf via bcftools call
}
