% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{grok_vcf}
\alias{grok_vcf}
\title{grok_vcf}
\usage{
grok_vcf(
  x,
  label = NA,
  keep.modifier = TRUE,
  long = FALSE,
  oneliner = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{GRanges input}

\item{label}{character value that will be added as a $label column}

\item{keep.modifier}{logical flag whether to keep modifier output}

\item{long}{logical flag whether to expand output so that every annotation has its own row, warning: will create duplicate rows for a given variant}

\item{oneliner}{logical flag whether to add a oneliner for the annotation}

\item{verbose}{logical flag whether to pritn output}
}
\description{
Does additional processing of annotated  vcf output and produces
a more readable granges output.

Warning: this function expects SnpEFF VCF input!!!
}
