% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{gsea_leading_edge}
\alias{gsea_leading_edge}
\title{gsea_leading_edge}
\usage{
gsea_leading_edge(gene.set, sig, draw.plot = T, cex = 1, asp = 2,
  eps = 1e-16, name = "")
}
\description{
Draws gsea plot for an input gene set and outputs leading edge
}
\author{
Marcin Imielinski
}

