% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{headf}
\alias{headf}
\title{headf}
\usage{
headf(x, n = 5, grep = NULL)
}
\arguments{
\item{x}{vector of iles}

\item{grep}{string to grep in files (=NULL)}
}
\description{
"head" +/- grep vector of files
}
\author{
Marcin Imielinski
}

