% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{hets}
\alias{hets}
\title{Simple het "caller" meant to be used at validated het SNP sites for tumor / normal pairs}
\usage{
hets(
  tum.bam,
  norm.bam = NULL,
  out.file,
  vcf.file,
  chunk.size1 = 1000,
  chunk.size2 = 100,
  mc.cores = 1,
  verbose = TRUE,
  na.rm = TRUE,
  filt.norm = TRUE
)
}
\arguments{
\item{tum.bam}{string path to tumor sample, input to Bamfile()}

\item{norm.bam}{string path to normal sample, input to Bamfile()(optional) (default = NULL)}

\item{out.file}{string path to TSV output file to be generated}

\item{vcf.file}{string path to VCF file of sites (eg hapmap or 1000G) at which to compute read counts}

\item{chunk.size1}{integer Number of variants to process from VCF file at a time (default = 1e3)}

\item{chunk.size2}{integer Number of variants to access from BAM file in a single iteration (default = 1e2)}

\item{mc.cores}{integer Number of cores in mclapply (default = 1)}

\item{verbose}{boolean Flag to increase verbosity (default = TRUE)}

\item{na.rm}{logical Flag to remove rows with NA counts (default = TRUE)}

\item{filt.norm}{logical Flag to remove any sites that have allele fraction of 0 or 1 or NA in MAF; if TRUE will remove any sites that have allele fraction 0 or 1 or NA in MAF}
}
\value{
nil
}
\description{
hets() outputs a tsv file of ALT ($alt.count.t, $alt.count.n) and REF ($ref.count.t,, $ref.count.n) read counts to out.file
for a tumor / normal pair across a set of sites specified by an input VCF
}
\author{
Marcin Imielinski
}
