% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{html_link}
\alias{html_link}
\title{html_link}
\usage{
html_link(href, text = NULL)
}
\arguments{
\item{href}{character vector of paths to link}

\item{text}{text to display}
}
\value{
character vector of html link text
}
\description{
returns text with html link
}
\author{
Marcin Imielinski
}
