% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{html_tag}
\alias{html_tag}
\title{html_tag}
\usage{
html_tag(tag, text = NULL, collapse = " ", ...)
}
\arguments{
\item{tag}{character vector of tags (without brackets)}

\item{text}{text to put inside tags}

\item{collapse}{how to collapse tags (=newline)}
}
\value{
character vector of html
}
\description{
makes a open and close html tag with optional text inside and optional (named) vector of name=value pairs contained inside of opening tag
}
\author{
Marcin Imielinski
}

