% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{igraph2graph}
\alias{igraph2graph}
\title{igraph2graph}
\usage{
igraph2graph(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
graph object
}
\description{
Converts igraph object into janky graphNEL object (for visualization in cytoscape)
and populates all edge features both via the edgeL and as NodeAttributes for visualization
in cytoscape
}
\author{
Marcin Imielinski
}

