% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{igv.loci}
\alias{igv.loci}
\title{igv.loci

Wrapper for igv function to dump table + screenshots for individual GRanges loci
that have sample column (default Tumor_Sample_Barcode) that is a key into
ind data.table where columns matching col.string are fetched and plotted

IGV host and port are taken from environment variables}
\usage{
igv.loci(
  mut,
  ind,
  out.path,
  sample.key = "Tumor_Sample_Barcode",
  sleep = 30,
  window = 400,
  host = Sys.getenv("IGV_HOST"),
  port = Sys.getenv("IGV_PORT"),
  overwrite = FALSE,
  snapshots = TRUE,
  verbose = TRUE
)
}
\description{
Wrapper for igv function to visualize a given (single) mutation in a maf file
}
\author{
Marcin Imielinski
}
