% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{igv}
\alias{igv}
\title{igv}
\usage{
igv(paths = NULL, gr = NULL, loci = NULL, snapshot = NULL,
  new = FALSE, reset = FALSE, wkspace = "PanLungWGS",
  host = Sys.getenv("IGV_HOST"), mac = !grepl("(^cga)|(node\\\\d+)", host),
  sort.locus = NULL, gsub.paths = list(c("/seq/picard_aggregation/",
  "/Volumes/seq_picard_aggregation/"), c("/xchip/singtex/",
  "/Volumes/xchip_singtex/"), c("/cga/meyerson/", "/Volumes/cga_meyerson/"),
  c("/seq/tier3b/", "/Volumes/seq_tier3b/"), c("/xchip/cle/",
  "/Volumes/xchip_cle/"), c("/xchip/cga/", "/Volumes/xchip_cga/"),
  c("/xchip/beroukhimlab/", "/Volumes/xchip_beroukhimlab/"), c("/cgaext/tcga/",
  "/Volumes/cgaext_tcga/"), c("~/", "~/home/"), c(Sys.getenv("HOME"),
  "~/home")), port = Sys.getenv("IGV_PORT"))
}
\arguments{
\item{paths}{file paths to display on current igv session (=NULL)}

\item{gr}{GRanges or GRAngesList of numeric genomic data or interval genomic annotations to send to IGV session, if gr has field $score then data will be dumped to .bw otherwise to .bed or .gff (=NULL)}

\item{loci}{GRanges or IGV parsable string specifying what window(s) on genome to view (=NULL)}

\item{snapshot}{file path to store snapshot in (has to be interpretable on file system where IGV is running)}

\item{new}{logical flag whether to start new IGV session}

\item{reset}{logical flag whether to reset connection between R and IGV (useful if IGV non responsive)}

\item{host}{character specifying host where IGV is running}

\item{mac}{logical flag specifying whether host is a local "mac" (ie then apply gsub.paths) otherwise keep paths the same}

\item{gsub.paths}{list of length 2 vectors specifying gsub args to apply to filenames when mac = TRUE}

\item{port}{integer specifying port where IGV is running}
}
\description{
Controls IGV on localhost (or specified host, separate from where R session is running).  Igv application must be running and listening to a specified port.  Then if you configure this port
via environment variables (IGV_HOST, IGV_PORT) in the current R session then you can apply the following usages

igv(fn) ## sends any given file(s) into igv (eg .bam, .wig, .bed)
igv(loci = cool.loci) ## plots the windows specified as GRanges or IGV-parsable strings (eg gene name)
igv(gr = cool.gr) ## sends granges object to igv session, Note: currently requires the ability to write to a public_html that is web viewable by computer running IGV
igv(snapshot = fn) ## sends current screen to file
igv(new = TRUE) ## refreshes current session
igv(reset = TRUE) ## resets connections, sometimes useful if IGV not responding

If alternate file paths are present on server (where R is runinng) and computer running IGV, then can specify gsub.paths variable which is a list of length 2 vectors
specifying how to convert file paths from arguments given to igv to ones that can be loaded locally.
}
\author{
Marcin Imielinski
}

