% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{img_link}
\alias{img_link}
\title{img_link}
\usage{
img_link(file, caption = NULL, embed = F, ...)
}
\arguments{
\item{file}{vector of (relative) image file paths to link to}

\item{caption}{character vector of captions to add (= '')}

\item{embed}{logical flag whether to imbed images instead of returning links (=FALSE)}

\item{...}{additional parameters to embed in tag (e.g. height and width)}
}
\value{
character vector of links (<a> tags) or image tags (<img> or <embed) to dump into an html document
}
\description{
Returns vector of html image links to files "file" with text "caption"

if embed = T, then will make img link, and additional arguments may be supplied to image tag (eg height, width)
}
\author{
Marcin Imielinski
}

