% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{ind2sub}
\alias{ind2sub}
\title{ind2sub}
\usage{
ind2sub(dim, ind, byrow = F)
}
\arguments{
\item{dim}{dimensions of matrix to query}

\item{ind}{1D index}

\item{byrow}{whether to calculate indices by row or column (= FALSE)}
}
\value{
length(ind) x 2 matrix of row and column index pairs corresponding to input ind in dim "dim" matrix
}
\description{
MATLAB style ind2sub function in R physical essence.  Provides the 2D row / column index for a
given 1D query
}
\author{
Marcin Imielinski
}
