% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{install.packages.bioc}
\alias{install.packages.bioc}
\title{install.packages.bioc}
\usage{
install.packages.bioc(pkg)
}
\arguments{
\item{pkg}{character vector of package names to install}
}
\description{
shortcut to install bioconductor packages
}
\author{
Marcin Imielinski
}
