% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{install.packages.github}
\alias{install.packages.github}
\title{install.packages.github}
\usage{
install.packages.github(pkg, username, branch)
}
\arguments{
\item{pkg}{character vector of package names to install}
}
\description{
shortcut to install github packages
}
\author{
Marcin Imielinski
}

