% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{is.dup}
\alias{is.dup}
\title{is.dup}
\usage{
is.dup(x)
}
\arguments{
\item{x}{vector or matrix to check}
}
\value{
logical vector of length(x) or nrow(x)
}
\description{
labels which vectors in x are part of a dup
returns logical TRUE if vector is part of a dup

Note: this is a twist on "duplicated" which only returns TRUE if a given element is a duplicate (i.e. duplicated ()  is FALSE
for the original version for the duplicate, while is.dup() will be TRUE for that element)

x can be vector or matrix
}
\author{
Marcin Imielinski
}
