% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{jhom}
\alias{jhom}
\title{wrapper around GxG homeology for junctions}
\usage{
jhom(
  event,
  pad = 100,
  thresh = 2,
  stride = 1,
  pad2 = 5,
  flip = FALSE,
  stats = FALSE,
  mc.cores = 1,
  anchor = TRUE,
  mat = FALSE
)
}
\arguments{
\item{event}{Junction object around which to compute homeology}

\item{pad}{padding to put around breakpoint window}

\item{thresh}{maximum string distance threshold for calling a bin homeologous}

\item{pad2}{pad to put around an output bin in order to measure homeology}

\item{flip}{flag whether to measure homeology with bins and their reverse complements}

\item{stats}{flag whether to return stats or gMatrix (default)}

\item{mc.cores}{integer number of cores [1]}

\item{anchor}{logical flag whether to return homeology results around native (inputted) coordinates or Anchored coordinates [TRUE]}

\item{mat}{logical whether to return results as a mat [FALSE] or gMatrix}
}
\value{
list of gMatrix objects (if stats = FALSE) and data.table of homeology stats (if stats = TRUE)
}
\description{
wrapper around GxG homeology for junctions
}
\author{
Marcin Imielinski
}
