% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{junction.support}
\alias{junction.support}
\title{junction.support}
\usage{
junction.support(
  reads,
  junctions = NULL,
  bwa = NULL,
  ref = NULL,
  pad = 500,
  pad.ref = pad * 20,
  realign = TRUE,
  walks = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{reads}{GRanges in SAM / BAM format e.g. output of read.bam or BWA, with fields $qname, $cigar, $flag $seq all populated in standard fashion, and optionally $AS}

\item{junctions}{Junction object}

\item{bwa}{RSeqLib BWA object and path to fasta file corresponding to the reference}

\item{ref}{optional DNAStringSet corresponding to reference genome sequence}

\item{pad}{padding around the junction breakpoint around  which to analyze contig and reference sequences, this should be several standard deviations above the average insert size (2000)}

\item{realign}{flag whether to realign or just use existing alignments}

\item{verbose}{logical flag (TRUE)}

\item{...}{additional parameters to contig support}
}
\value{
reads re-aligned to the reference through the contigs with additional metadata describing features of the alignment
}
\description{
Takes as input a GRanges of bam alignments (e.g. outputted from bamUtils::read.bam) and a GRanges of rearranged
reference aligned contigs (e.g. output of RSeqLib::BWA) and a set of Junction objects, and outputs reads supporting
these junctions by building a contig around each junction (from the reference) and then running contig.support (see
that functions docuemntation for criteria)
}
\author{
Marcin Imielinski
}
