% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{list.expr}
\alias{list.expr}
\title{list.expr}
\usage{
list.expr(x)
}
\arguments{
\item{x}{input vector}
}
\value{
character vector of command to create the input vector
}
\description{
Takes a character or numeric vector and makes an expression for re-creating that character in source code
}
\author{
Marcin Imielinski
}

