% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{maf2vcf}
\alias{maf2vcf}
\title{maf2vcf}
\usage{
maf2vcf(maf, fn)
}
\arguments{
\item{maf}{maf data.frame}

\item{fn}{output file}
}
\description{
Dumps maf data frame to VCF file "fn"
}
\author{
Marcin Imielinski
}

