% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{maf_classify}
\alias{maf_classify}
\title{maf_classify}
\usage{
maf_classify(maf)
}
\arguments{
\item{maf}{maf data.frame}
}
\value{
variant categories
}
\description{
Re-classifies oncotated variants
}
\details{

}
\author{
Marcin Imielinski
}

