% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{maf_coding}
\alias{maf_coding}
\title{maf_coding}
\usage{
maf_coding(maf, inclusive = T)
}
\arguments{
\item{maf}{maf data.frame}
}
\value{
logical vector specifying whether row satisfies the criterion
}
\description{
Scans "Variant_Classification" field in maf and outputs TRUE if variant is in protein coding region (includes synonymous)
}
\author{
Marcin Imielinski
}

