% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{maf_disp}
\alias{maf_disp}
\title{maf_disp}
\usage{
maf_disp(maf, flavor = NULL, sorted = F, show.pat = TRUE,
  extra_cols = NULL, gene = NULL, pat = NULL)
}
\arguments{
\item{maf}{data.frame with Oncotated maf columns}

\item{flavor}{character specifying 'functional' or 'validation' flavors, which correspond to special column slices of maf data.frame (= NULL)}

\item{sorted}{logical flag whether to return output sorted by gene, variant classificaiton, uniprot site, then patient (=FALSE)}

\item{show.pat}{logical flag whether to show patient (=TRUE)}

\item{gene}{character vector of Hugo_Symbol to subset on (=NULL)}

\item{pat}{character vector of Tumor_Sample_Barcodes to subset on (=NULL)}

\item{extra_col}{character vector of additional columns to include (=NULL)}
}
\value{
character vector or sliced data.frame
}
\description{
Quick display of rows of data.frame holding contents of Oncotated maf file
}
\author{
Marcin Imielinski
}

