% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{maf_ns}
\alias{maf_ns}
\title{maf_ns}
\usage{
maf_ns(maf, inclusive = T)
}
\arguments{
\item{maf}{maf data.frame}
}
\value{
logical vector specifying whether maf row satisfies the criterion
}
\description{
Scans "Variant_Classification" field in maf and outputs TRUE if variant is non-synonymous
}
\author{
Marcin Imielinski
}

