% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{maflite}
\alias{maflite}
\title{maflite}
\usage{
maflite(maf)
}
\arguments{
\item{maf}{maf.data.frame}
}
\value{
data.frame in maf.lite format
}
\description{
take maf data frame and returns columns corresponding to "maflite" format
https://confluence.broadinstitute.org/display/CGA/Onctoator#Oncotator-Mafliteformat
}
\author{
Marcin Imielinski
}

