% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{match.seq}
\alias{match.seq}
\title{match.seq}
\usage{
match.seq(query, subject, mc.cores = 1, verbose = FALSE, ...)
}
\arguments{
\item{query}{character or DNAStringSet}

\item{subject}{character or DNAStringSet}

\item{mc.cores}{multithreading for parsing}

\item{...}{additional params to PDict}
}
\description{
(Exact) matches a set of character query sequences against a set of (optionally named) subject sequences
Returning a GRanges seqnames and coordinates with query.id (and optionally query name) as meta data.
}
