% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{morder}
\alias{morder}
\title{morder}
\usage{
morder(A, orient = 1)
}
\arguments{
\item{A}{matrix of values}

\item{orient}{integer orientation, if 1 will do row-wise ordering, otherwise column ordering (=1)}
}
\value{
input matrix with rows and columns ordered
}
\description{
matrix order wrt columns ..
ie ordering rows matrix based on left to right ordering of columns (if MARGIN = 1)
OR  ordering columns of  matrix based on top to bottom ordering of rows (if MARGIN = 2)
}
\author{
Marcin Imielinski
}

