% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{mski_alpha}
\alias{mski_alpha}
\title{mski_alpha}
\usage{
mski_alpha(col, alpha = FALSE)
}
\arguments{
\item{col}{string Any of the three kinds of R color specifications, i.e., either a color name (as listed by colors()), a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb), or a positive integer i meaning palette()[i]}

\item{alpha}{boolean indicating whether the alpha channel (opacity) values should be returned (default = FALSE)}
}
\value{
string
}
\description{
Wrapper combining 'col2rgb()' and 'rgb()' for single colors

Originally called 'alpha()' in 'mskilab/bamUtils'
}
