% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{mtable}
\alias{mtable}
\title{mtable}
\usage{
mtable(mat)
}
\arguments{
\item{mat}{input matrix}
}
\value{
unique rows of mat, with additional column $count on the left
}
\description{
tabulates unique rows values for matrix / data frame
}
\author{
Marcin Imielinski
}

