% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{mut_genecluster}
\alias{mut_genecluster}
\title{mut_genecluster}
\usage{
mut_genecluster(genes, maf, cov, p.thresh = 0.05, min.cluster.size = 2,
  bonferonni = TRUE)
}
\description{
Greedy divisive clustering of genes based on mutation rates along a provided order (eg order of gene expression)

Outputs a list of gene clusters (each list a character vector of gene symbols)
}
\author{
Marcin Imielinski
}
