% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{mutclusters}
\alias{mutclusters}
\title{mutclusters}
\usage{
mutclusters(maf, d = 0, k = 1, method = "single", max.cluster = TRUE)
}
\arguments{
\item{maf}{maf data.frame}

\item{d}{max distance threshold in amino acid space}

\item{k}{minimum number of mutations in returned clusters}

\item{method}{character specifying "single" or "complete" linkage clustering of mutations}

\item{max.cluster}{logical flag whether to return the gene with the largest cluster (if TRUE) or the most number of clusters (if FALSE) (=TRUE)}
}
\value{
genes ranked by numbers of cluster or max.cluster size
}
\description{
Returns genes with a degree of mutaiton clustering (e.g. ranked by how many k>k.thresh clusters with d<d.thresh pairwise distance, or
the largest cluster with those characteristics)

if max.cluster = TRUE returns maximum size cluster in gene where either all (method == complete) or at least one (method single)
mutation pair is within distance d

if max.cluster = F then returns number of clusters of mutations of count greater than k within a distance d per gene

eg d = 0, k = 1, will give the number of unique sites with more than 1 perfectly recurrent mutation per gene

Clustering is by default using single-linkage agglomerative clustering, but any method that is input to hclust can be used
}
\author{
Marcin Imielinski
}

