% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{mutpairsd}
\alias{mutpairsd}
\title{mutpairsd}
\usage{
mutpairsd(maf, d = 0)
}
\arguments{
\item{maf}{maf data.frame}

\item{d}{distance threshold in amino acid space}
}
\value{
clusters of mutations with how many pairs of variants supporting
}
\description{
Takes maf data.frame and outputs table that lists how many pairs there are <= distance d in amino acid space
}
\author{
Marcin Imielinski
}

