% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{mutrate_window}
\alias{mutrate_window}
\title{mutrate_window}
\usage{
mutrate_window(genes, maf, cov, window = 100)
}
\description{
Computes mutation rates along k gene "windows" along an ordered list "genes" of genes.
}
\author{
Marcin Imielinski
}
