% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{nnnusvr}
\alias{nnnusvr}
\title{nnnusvr}
\usage{
nnnusvr(X, y, nu = 0.85, C = 1, nn = TRUE, intercept = FALSE)
}
\arguments{
\item{X}{numeric matrix of n data points by m features}

\item{y}{numeric length(n) response vector}

\item{nu}{scalar numeric between 0 and 1 representing fraction of data to use as support vectors}

\item{C}{scalar non-negative numeric representing the cost}

\item{nn}{logical scalar whether to constrain the output weights to be non-negative}
}
\description{
Implementation of non-negative linear nu-support vector regression using CPLEX, given
length n response vector y and n x m data matrix X returns a length m
weight vector w linear SV regresssion result (or length m+1 result if
intercept = TRUE is specified).

Additional feature nn (default = TRUE) allows constraining the weight vector to be non-negative
}
\author{
Marcin Imielinski
}
