% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{nona}
\alias{nona}
\title{nona}
\usage{
nona(dt, thresh = 0.1)
}
\value{
data.table with a subset of columns that have at least a given fraction of non NA entries
}
\description{
Given data.frame / data.table dt outputs only the columns that are non.na 
in at least thresh fraction of the rows
}
\author{
Marcin Imielinski
}
