% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{nz}
\alias{nz}
\title{nz}
\usage{
nz(x, zero = 0, full = FALSE, matrix = TRUE)
}
\arguments{
\item{x}{length(x)}

\item{zero}{integer specifying what to use as the "zero" value in the input (=0)}
}
\value{
data.frame of row id col id value pairs
}
\description{
outputs the nonzero entries of a vector or array
}
\author{
Marcin Imielinski
}
