% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{oncoprint}
\alias{oncoprint}
\title{oncoprint}
\usage{
oncoprint(
  tumors = NULL,
  oncotab = NULL,
  genes = c("KRAS", "EGFR", "BRAF", "TP53", "TERT", "CCND1", "MYC", "PIK3CA", "PTEN",
    "CDKN2A", "ARID1A", "SMARCA4"),
  split = NULL,
  sort = TRUE,
  sort.genes = sort,
  sort.tumors = sort,
  columns = NULL,
  noncoding = FALSE,
  cna = TRUE,
  tmb = TRUE,
  pp = TRUE,
  signature = TRUE,
  svevents = TRUE,
  basic = FALSE,
  ppdf = TRUE,
  return.oncotab = FALSE,
  return.mat = FALSE,
  wes = TRUE,
  drop = TRUE,
  track.height = 1,
  signature.thresh = 0.2,
  signature.main = c(1:5, 7, 9, 13),
  track.gap = track.height/2,
  split.gap = 1,
  colnames.fontsize = 10,
  rownames.fontsize = 10,
  track.fontsize = 10,
  mc.cores = 1,
  verbose = FALSE,
  height = 20,
  width = 20,
  ...
)
}
\arguments{
\item{tumors}{keyed table of tumors (aka pairs table) with field $oncotable which points to a cached .rds file of an oncotable e.g. produced by oncotable function or Oncotable module / task}

\item{oncotab}{output from oncotable function with field $id}

\item{genes}{character vector of genes}

\item{split}{character of name of column in tumors table to split on (NULL)}

\item{sort}{logical flag whether to automatically sort rows i.e. genes and columns i.e. tumors in a "stair step" pattern or default to the provided (TRUE)}

\item{sort.genes}{logical flag whether to sort rows i.e. genes with respect to their frequency (TRUE)}

\item{sort.tumors}{logical flag whether to sort columns i.e. patients in a stairstep pattern with respect to the provided gene order (TRUE)}

\item{columns}{additional columns of tumors matrix to plot as horizontal tracks below the main track}

\item{noncoding}{logical flag whether to show non protein coding mutations}

\item{pp}{logical flag whether to show purity / ploidy (if data is provided / available) (TRUE)}

\item{svevents}{logical flag whether to show events (if data is provided / available) (TRUE)
o=@param tmb logical flag whether to show TMB bar plot (TRUE)}

\item{ppdf}{whether to print to pdf via ppdf}

\item{return.mat}{whether to return.mat}

\item{wes}{logical flag whether to use wesanderson coolors}

\item{track.height}{height of tracks in 'cm'}

\item{signature.thresh}{lower threshold for non main signature fraction in at least one sample to plot}

\item{signature.main}{integer indices of main COSMIC signatures to keep}

\item{split.gap}{gap between splits}

\item{mc.cores}{multicore threads to use for $oncotable loading from tumors table (not relevant if oncotab provided)}

\item{...}{other arguments to ppdf}

\item{sv.stack}{logical flag whether to stack bar plot simple and complex SV event counts (FALSE)}

\item{signatures}{logical flag whether to show signatures (if data is provided / available) (TRUE)}

\item{tmb.log}{logical flag whether to log TMB + 1 (TRUE)}

\item{cex}{length 1 or 2 vector canvas expansion factor to apply to the oncoprint itself (relative to 10 x 10 cm) (c(1,3))}
}
\value{
ComplexHeatmap object (if ppdf = FALSE), and genotype matrix (if return)
}
\description{
Simple wrapper around to oncoPrint from complexHeatmap package to allow quick plotting
of patients x genes + metadata.  Uses the data gathered by oncotab to generate quick simple
plots that include a core matrix of genes x patients containing data on  SCNA, (complex), fusions,
SNV, and indels.  Additional tracks plotting log TMB + 1, log SV burden, complex events,
SNV signatures can be provided.
}
\author{
Marcin Imielinski
}
