% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{oncotable}
\alias{oncotable}
\title{oncotable}
\usage{
oncotable(
  tumors,
  gencode = NULL,
  verbose = TRUE,
  amp.thresh = 4,
  filter = "PASS",
  del.thresh = 0.5,
  mc.cores = 1
)
}
\arguments{
\item{tumors}{keyed data.table i.e. keyed by unique tumor id with specific columns corresponding to  paths to pipeline outputs(see description)}

\item{gencode}{path to gencode .gtf or .rds with GRanges object, or a GRanges object i.e. resulting from importing the (appropriate) GENCODE .gtf via rtracklayer, note: this input is only used in CNA to gene mapping}

\item{verbose}{logical flag}

\item{amp.thresh}{SCNA amplification threshold to call an amp as a function of ploidy (4)}

\item{del.thresh}{SCNA deletion threshold for (het) del as a function of ploidy (by default cn = 1 will be called del, but this allows additoinal regions in high ploidy tumors to be considered het dels)}

\item{mc.cores}{number of cores for multithreading}
}
\description{
Takes as input (keyed) "tumors" (aka pairs) table which a metadata table with specific
columns pointing to paths corresponding to one or more of the following pipeline outputs:

$annotated_bcf  Path to annotated.bcf file that is the primary output of SnpEff module from which TMB and basic mutation
descriptions are extracted along with their basic categories (these will comprising the core of the oncoplot are computed)

$fusions  Path to fusion.rds file that is the primary output of the Fusions modjle, from which protein coding fusions will
be computed for

$jabba_rds  Path to jabba.simple.rds output representing primary output of JaBbA module from which SCNA and overall
junction burden are computed

$complex    Path to complex.rds gGnome cached object that is the primary output of Events module, from which simple
and complex event burdens are computed

$signature_counts Path to signature_counts.txt that is the primary output of Signatures module from which SNV signature
counts are computed

The function then outputs a melted data.table of "interesting" features that can be saved and/or immediately output
into oncoprint.  This data.table will at the very least have fields $id $type (event type), $track, and  $source
populated in addition to a few other data type specific columns.

The $source column is the name of the column of tumors from which that data was extracted, and track is a grouping
variable that allows separation of the various data types. 

All the paths above can be NA or non existent, in which case a dummy row is inserted into the table so that downstream
applications know that data is missing for that sample.
}
\author{
Marcin Imielinski
}
