% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{oneoffs}
\alias{oneoffs}
\title{Calls samtools mpileup to dump tsv of "one off" variants / sites}
\usage{
oneoffs(
  out.file,
  bam,
  ref,
  min.bq = 30,
  min.mq = 60,
  indel = FALSE,
  chunksize = 10000,
  verbose = TRUE
)
}
\arguments{
\item{out.file}{string Path to file in which to dump tsv}

\item{bam}{string Path to BAM file}

\item{ref}{tring Path to reference FASTA}

\item{min.bq}{integer Minimum base quality}

\item{min.mq}{integer Minimum mapping quality}

\item{indel}{boolean Flag whether to collect one off indels (default is substitution)}

\item{chunksize}{integer Number of mpileup lines to put into memory}

\item{verbose}{boolean Flag to increase verbosity (default = TRUE)}
}
\description{
Calls samtools mpileup to dump tsv of "one off" variants / sites (i.e. that are present in exactly one read per site)
}
\note{
The denominator (i.e. total reads) is just the sum of counts$records
}
