% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{pad}
\alias{pad}
\title{pad}
\usage{
pad(x, k, clip = T)
}
\arguments{
\item{x}{integer vector to pad}

\item{k}{window around each entry to pad}

\item{clip}{logical flag whether to clip elements below 0 (=TRUE)}
}
\value{
"padded" integer vector of unique entires with entries in k window around each input included
}
\description{
pads an (integer) vector with k places below and above its lowest and highest value

(by default, clips at 0)

useful for querying around specific entires of vector, matrix, data.frame, GRanges ewtc
}
\author{
Marcin Imielinski
}

