% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{parsesnpeff}
\alias{parsesnpeff}
\title{parsesnpeff}
\usage{
parsesnpeff(vcf, id)
}
\arguments{
\item{vcf}{path to vcf}
}
\value{
GRanges object of all variants and annotations
}
\description{
parses vcf file containing SnpEff annotations on Strelka calls
}
\author{
Kevin Hadi
}

