% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{pindel}
\alias{pindel}
\title{pindel}
\usage{
pindel(outdir, bams, intervals = NULL, isizes = NULL,
  hg = Sys.getenv("GATK.FASTA"), genome = "hg19", outroot = "out",
  run = F, verbose = T, write_bam = F, oncotate = F, threads = 1,
  window.size = 10, other.args = "")
}
\arguments{
\item{outdir}{out directory to output to}

\item{bams}{vector of input bams}

\item{intervals}{GRanges of to run on (=NULL)}

\item{isizes}{integer insert.size to use}

\item{hg}{genome fasta location (=Sys.getenv('GATK.HG19'))}

\item{genome}{genome build (='hg19')}

\item{outroot}{prefix to give output files (='out')}

\item{run}{logical flag whether to run immediately or just return character vector of command (= FALSE)}

\item{verbose}{logical flag}

\item{write_bam}{logical flag whether to write the bam (=FALSE)}

\item{oncotate}{logical flag whether to oncotate output into MAF files (=TRUE)}

\item{threads}{number of threads to use}

\item{window.size}{integer window size to use (=10)}

\item{other.args}{other args to add (='')}
}
\value{
character vector of command(s) (only if run = FALSE), otherwise just runs command with system call.
}
\description{
calls pindel on a set of input bams and a given set of targets,
outputting to target.dir

intervals are given as GRanges
}
\author{
Marcin Imielinski
}
