% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{plop}
\alias{plop}
\title{plop}
\usage{
plop(fn, prefix = NULL, force = NULL)
}
\arguments{
\item{fn}{character vector of filenames to "plop" into ~public_html/prefix (Sys.getenv('PLOP.DIR') is used as alternative if defined)}

\item{prefix}{character prefix to add to filenames after plopping (can include subdirectories which can be inspected)}
}
\description{
grabs file and plops into public_html (or Sys.getenv('PLOP.DIR') if defined)

prefix will be added to left of file name (can include firectory)

if fn is list then prefix is expanded to unlisted fn

Useful for inspecting a specific subset of analysis files eg when debugging.
}
\author{
Marcin Imielinski
}

