% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{plot_multiplicity}
\alias{plot_multiplicity}
\title{plot_multiplicity}
\usage{
plot_multiplicity(individual, maf, plot.reads = F)
}
\description{
Plots allele fractions and colors in an individual  based on multiplicity call stratified by total copy number value
"individual" is a list or data frame with field $Individual_Id, $Purity

maf is a maf file annotated with multiplicity (ie the output of allele_multiplicity) with additional fields
$cn.tot, $mult, $mult.p
}
\author{
Marcin Imielinski
}

