% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{pmGSEA}
\alias{pmGSEA}
\title{poor mans GSEA}
\usage{
pmGSEA(gene.sets, sig.table, min.perms = 100, max.perms = 1e+05,
  positive.selection = T, length.filter = F, length.range = c(5, 50),
  mc.cores = 1, plot.hist = F, verbose = F, bootstrap = T,
  rank.test = F)
}
\arguments{
\item{gene.sets}{a named list of character vectors, each list item is a gene set, i.e. a character vector of genes}

\item{sig.table}{named vector of p values from an analysis e.g. mutSig, the names of the genes are}

\item{min.perms}{minimum number of permutations to do in the adaptive permutation test}

\item{max.perms}{maximum number of permutations to do in the adaptive permutation test}

\item{length.range}{length 2 integer vector specifying min and max gene set size to score after intersection with genes in sig.table default: c(5,50)}
}
\description{
pmGSEA "poor man's GSEA " ***

Given a gene.set (character vector) or gene.sets (list of character vectors)
and given a named vector of significance values or table of significant genes (sig.table)
(if table then significance column is $p or first column) identifies gene sets that have significant
negative deviation of a "signed K-S" statistic vs uniform distribution  (ie have p values significantly
clustering towards zero) ie are significantly enriched in genes showing positive selection.

if positive.selection = F, will identify sets with significantly positive deviation of a "signed K-S" statistic (ie have p values significantly clustering towards 1)
these are sets showig significant negative selection.

All p-values are computed against a distribution of signed K-S statistic obtained through permutation using random gene sets of the same size chosen from sig.table

Will adaptively perform permutations between minperms and maxperms using following rule of thumb: if there are <PERM.THRESH permutations with
greater than (lower.tail = F) or less than (lower.tail = T) score than observed score, then will compute additional perms

*** actually not much poorer than the original GSEA, basically a reimplementation of Mootha et al Nat Gen 2002
}
\author{
Marcin Imielinski
}
