% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{pp_plot}
\alias{pp_plot}
\title{pp_plot}
\usage{
pp_plot(
  jabba_rds = NULL,
  cov.fname = NULL,
  hets.fname = NULL,
  allele = FALSE,
  field = NULL,
  plot.min = -2,
  plot.max = 2,
  scatter = FALSE,
  bins = 500,
  height = 800,
  width = 800,
  output.fname = "./plot.png",
  save = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{jabba_rds}{(character) JaBbA output}

\item{cov.fname}{(character) path to coverage GRanges (supply if allele = FALSE)}

\item{hets.fname}{(character) path to sites.txt (supply if allele = TRUE)}

\item{allele}{(logical) allelic CN? default FALSE}

\item{field}{(character) default ratio if allele is FALSE and count if allele is TRUE}

\item{plot.min}{(numeric) minimum CN default -2}

\item{plot.max}{(numeric) max CN (factor times ploidy) default 2}

\item{scatter}{(logical) default FALSE}

\item{bins}{(numeric) number of histogram bins default 500}

\item{height}{(numeric) plot height}

\item{width}{(numeric) plot width}

\item{output.fname}{(character) path of output directory}

\item{verbose}{(logical)}
}
\value{
output.fname
}
\description{
pp_plot
}
\details{
create histogram of estimated copy number given purity and ploidy
}
\author{
Zi-Ning Choo
}
