% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{ppdf}
\alias{ppdf}
\title{ppdf}
\usage{
ppdf(expr, filename = "plot.pdf", height = 10, width = 10, cex = 1,
  title = NULL, byrow = TRUE, dim = NULL, cex.title = 1, ...)
}
\arguments{
\item{expr}{Plotting expression eg plot(runif(1000), runif(1000))}

\item{filename}{filename under ~/public_html/ or Sys.getenv('PPDF.DIR') to dump plots to (='plot.pdf')}

\item{height}{integer pixel height of plot (=1000)}

\item{width}{integer pixel width of plot (=1000)}

\item{cex}{expansion factor of plot from "default size" either length 1 scalar or length 2 vector specifying height and width expansion (=1)}

\item{title}{title to add to plot (='')}

\item{dim}{length 2 integer vector, if expr contains multiple plot calls then will output to matrix of plots with specified dim (=NULL)}

\item{cex.title}{character expansion factor to title)}

\item{...}{additional arguments to pdf()}
}
\description{
sends quick plot to ~public_html/plot.pdf.  If PPDF.DIR env variable defined, then will send to that directory (i.e. instead of public_html)
Useful for doing quick standard plots to a static location which one views through tabs in Chrome or other web browser.
}
\author{
Marcin Imielinski
}

