% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{PPplots}
\alias{PPplots}
\title{PPplots}
\usage{
PPplots(
  cov,
  hets,
  pu,
  pl,
  xmax = 10,
  hist_breaks = 10000,
  outputdir = ".",
  prefix = "",
  suffix = "",
  N_subsample = 10000
)
}
\arguments{
\item{cov}{bin coverage depth}

\item{hets}{het_pileup}

\item{pu}{purity}

\item{pl}{ploidy}

\item{xmax}{maximum value for x-axis of histograms}

\item{outputdir}{output directory for plots}

\item{prefix}{a prefix to use for output plots}

\item{suffix}{a suffix to use for output plots}

\item{N_subsample}{How many entries to randomly sample from hets when generating the contour plot}
}
\description{
Plot bin and het copy number histogram as well as a contour plot for allele fractions.
}
\author{
Alon Shaiber
histogram of bin copy number
histogram of het copy numbers
density plots for hets
}
