% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{preprocess_cov_for_dryclean}
\alias{preprocess_cov_for_dryclean}
\title{preprocess_cov_for_dryclean}
\usage{
preprocess_cov_for_dryclean(
  cov,
  field = "reads.corrected",
  template_rds = "",
  seqnames.to.include = c(as.character(1:22), "X"),
  output.field = "reads.corrected",
 
    pon_dict = "/gpfs/commons/groups/imielinski_lab/DB/modules/dryclean/dryclean_pon_dictionary.tsv",
  pon_name = NULL,
  nochr = TRUE
)
}
\arguments{
\item{cov}{GRanges with coverage data}

\item{field}{field to use as input from the coverage GRanges}

\item{template_rds}{path to a RDS file containing the template GRanges to use for re-binning}

\item{seqnames.to.include}{which seqnames to include in the output}

\item{output.field}{field name for the output normalized and rebinned data}

\item{pon_dict}{path to the TAB delimited file PON dictionary (if no template_rds was provided then you must provide a pon_name)}

\item{pon_name}{name of the PON. This is used to get the path to the template RDS from the pon_dict (if no template_rds was provided then you must provide a pon_name)}

\item{nochr}{by default all chr prefix is removed, if you wish the output to have chr you must specify nochr=FALSE}
}
\value{
new.cov GRanges object with the normalized and rebinned coverage values
}
\description{
preprocess coverage file by rebinning to match the ranges that are compatible with the PON and by normalizing the coverage by the mean
}
\author{
Alon Shaiber
}
