% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{process_gencode}
\alias{process_gencode}
\title{process_gencode}
\usage{
process_gencode(gencode = NULL)
}
\arguments{
\item{gencode}{path to gencode file. Gencode file must be either rds or some format accepted by rtracklayer::import (e.g. GTF)}
}
\value{
gencode_gr GRanges
}
\description{
Helper script to process gencode parameter
}
\author{
Marcin Imielinski
}
