% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{pscatter}
\alias{pscatter}
\title{pscatter}
\usage{
pscatter(x, y, text = "", color = NULL, size = NULL, mode = "markers",
  type = "scatter")
}
\description{
Quick plotly scatterplot
}
\author{
Marcin Imielinski
}
