% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{qq_pval}
\alias{qq_pval}
\title{qq plot given input p values}
\usage{
qq_pval(obs, highlight = c(), exp = NULL, lwd = 1, bestfit = T,
  col = NULL, col.bg = "black", pch = 18, cex = 1, conf.lines = T,
  max = NULL, qvalues = NULL, label = NULL, subsample = NA, ...)
}
\arguments{
\item{obs}{vector of pvalues to plot, names of obs can be intepreted as labels}

\item{highlight}{optional arg specifying indices of data points to highlight (ie color red)}

\item{lwd}{integer, optional, specifying thickness of line fit to data}

\item{pch}{integer dot type for scatter plot}

\item{cex}{integer dot size for scatter plot}

\item{conf.lines}{logical, optional, whether to draw 95 percent confidence interval lines around x-y line}

\item{max}{numeric, optional, threshold to max the input p values}

\item{label}{character vector, optional specifying which data points to label (obs vector has to be named, for this to work)}

\item{samp}{integer, optional specifying how many samples to draw from input data (default NULL)}
}
\description{
qq plot given input p values
}
\author{
Marcin Imielinski, Eran Hodis
}

