% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{qq_repel}
\alias{qq_repel}
\title{QQ plot with repel}
\usage{
qq_repel(
  ps,
  label = rep("", length(ps)),
  conf.lines = FALSE,
  ci = 0.95,
  print = TRUE
)
}
\arguments{
\item{ps}{vector of p values}

\item{label}{length(ps) vector of labels}
}
\description{
QQ plot with repel
}
\author{
Ashley Doane
}
