% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{qstat}
\alias{qstat}
\title{qstat}
\usage{
qstat(full = FALSE, numslots = TRUE)
}
\description{
Tabulates cluster usage (qstat()) or if full = TRUE flag given will dump out
all running jobs in a data.table
}

