% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{query_lsf_out}
\alias{lsf_query}
\alias{query_lsf_out}
\title{query_lsf_out}
\usage{
lsf_query(dir = NULL, jname = NULL, detailed = F, mc.cores = 1)
}
\arguments{
\item{dir}{character specifyhing directory where .out and .err files are located}

\item{jname}{character vector names of jobs (as specified in bsub_cmd}

\item{detailed}{logical flag specifying whether to return "detailed" information}

\item{mc.cores}{integer specifying how many cores to use to parse the output data}
}
\value{
data.frame of job info
}
\description{
parses "out" and "err" files of jobs with jname root to identify exit status and error codes of jobs
}
\author{
Marcin Imielinski
}

