% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{quickSig}
\alias{quickSig}
\title{quickSig}
\usage{
quickSig(maf, cov, patients = NULL, genes = NULL, analyze.categories = F,
  remove.silent = TRUE, limit.cat = NULL, two.tailed = TRUE)
}
\description{
Quick implementation of Mike / Gaddy binomial / poisson model.  Requires category based coverage output (*.per_gene.coverage.txt file)
computed during mutsig preprocess step.
(run mutsig_preprocess setting the following additional flags:
  P.output_per_gene_coverage = true;
  P.output_per_gene_mutation_counts = true;
  P.simplified_gene_sample_coverage_table = false;
  P.simplified_gene_sample_mutation_counts_table = false;
)

Computes context-category specific mutation rates either across whole cohort or within strata of gene-patient categories.

Outputs a significance table with the following columns for each category (if analyze.categories = TRUE) and/or category "tot" which is across all categories
o.k = observed mutations of category k
e.k = expected mutations of category k given background model
eff.k = log(o.k / e.k) for category k
p.k = p value of deviation from expectation under poisson model
q.k = q value of deviation
}
\author{
Marcin Imielinski
}
