% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{qviz}
\alias{qviz}
\title{qviz}
\usage{
qviz(res = NULL, queue = NULL, field = "h_vmem", frac = FALSE,
  all = FALSE, plot = TRUE)
}
\description{
Plots resources (default h_vmem requestes) across cluster either for
provided slice of qstat(full = TRUE) output or a fresh call to qstat(full = TRUE)
}
\author{
Marcin Imielinski
}
