% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{read_pon_dict}
\alias{read_pon_dict}
\alias{read_pon_dict_path}
\title{read_pon_dict}
\usage{
read_pon_dict_path(pon_dict_path)
}
\arguments{
\item{pon_dict_path}{path to the TAB delimited file.}
}
\value{
pon_dict data.table object
}
\description{
Check the validity and read the TAB delimited dictionary for dryclean PONs
}
\author{
Alon Shaiber
}
