% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{readRDA}
\alias{readRDA}
\title{readRDA}
\usage{
readRDA(fn)
}
\arguments{
\item{fn}{file name of .rda or .RData file}
}
\value{
object containing all the elements of the environment stored in fn
}
\description{
loads Rdata environment into a list variable and returns
(to mirror RDS functionality)
}
\author{
Marcin Imielinski
}

