% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{reorder_maf}
\alias{reorder_maf}
\title{reorder_maf}
\usage{
reorder_maf(maf, include.other = TRUE)
}
\arguments{
\item{maf}{data.frame of MAF}

\item{include.other}{logical flag whether to include non-standard maf columns after the standard ones (=TRUE)}
}
\value{
data.frame representing MAF in standard column order
}
\description{
Re-orders maf to comply with the TCGA MAF specifications (v2.2), tacking on all "non-standard" columns after the initial 32
}
\author{
Marcin Imielinski
}

