% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{row.slice}
\alias{row.slice}
\alias{\%~\%}
\title{row.slice}
\usage{
df \%~\% ...
}
\arguments{
\item{df}{data.frame}

\item{regex}{string to match or number in which case that column is returned (same behavior for data.table)}
}
\value{
slices of data.frame matching regex
}
\description{
Hacked operator for subsetting rows of data.frames, DataFrames, data.tables, GRanges

df %~% 1:5 retrieves first five rows

df %~% NA retrieves 5 random rows
}
\author{
Marcin Imielinski
}
