% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{rr}
\alias{rr}
\title{rr}
\usage{
rr(x, y = c())
}
\arguments{
\item{x}{expression or variables}

\item{...}{additional variables}
}
\description{
select row of data.frame or data.frame- like object data.table
}
