% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{score.walks}
\alias{score.walks}
\title{score.walks}
\usage{
score.walks(
  wks,
  bam = NULL,
  reads = NULL,
  win = NULL,
  wins = NULL,
  use.discordant = FALSE,
  rthresh = 4,
  thresh = 1e+05,
  pad = 10000,
  raw = FALSE,
  allpaths = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{wks}{GRangesList of walks}

\item{bam}{bam file}

\item{reads}{GRanges of reads with BX tag}

\item{win}{genomic window in which to score (default is just reduce(unlist(wks))))}

\item{wins}{tiles to chop up genome further (beyond walk segments)}

\item{use.discordant}{logical flag whether to process discordant}

\item{raw}{returns raw barcode by walk matrix of barcode scores}
}
\value{
scores of walks or (if raw == tRUE) raw barcode to walk maps
}
\description{
Scores GRangesList of walks against GRanges of 10X reads with $BX tag.
}
\author{
Marcin Imielinski
}
