% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{seconds}
\alias{seconds}
\title{seconds}
\usage{
seconds(x)
}
\description{
Takes string representing time elapsed and returns numeric string representing number of seconds
represented by this string or NA
}
\author{
Marcin Imielinski
}
