% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{set.comp}
\alias{set.comp}
\title{set.comp}
\usage{
set.comp(s1, s2)
}
\arguments{
\item{s1}{vector corresponding to "set 1"}

\item{s2}{vector corresponding to "set 2"}
}
\value{
list with fields $left, $middle, and $right corresponding to vectors that are in the left setdiff, intersection, right setdiff respectively
}
\description{
Compares two sets and outputs data frame with "left", "middle", "right" members
}
\author{
Bryan Hernandez
}

