% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{sgrep}
\alias{sgrep}
\title{System grep}
\usage{
sgrep(paths, expr)
}
\arguments{
\item{paths}{character of file paths}

\item{expr}{regular expression to match}
}
\value{
data.table of all non
}
\description{
Greps for expression and returns list with hits, indexed by filename
}
\author{
Marcin Imielinski
}
