% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{sgrepl}
\alias{sgrepl}
\title{System grep logical}
\usage{
sgrepl(paths, expr)
}
\arguments{
\item{paths}{character of file paths}

\item{expr}{regular expression to match}
}
\value{
data.table of all non
}
\description{
Greps for expression and returns logical vector if match exists in file
}
\author{
Marcin Imielinski
}
