% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skitools-package.R
\docType{package}
\name{skitools-package}
\alias{skitools}
\alias{skitools-package}
\title{skitools: Various mskilab R utilties}
\description{
R miscellaneous utilities for basic data manipulation, debugging, viz, lsf management, and common mskilab tasks.
}
\author{
\strong{Maintainer}: Marcin Imielinski \email{mimielinski@nygenome.org}

}
