% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{slurm_dt}
\alias{slurm_dt}
\title{slurm_dt}
\usage{
slurm_dt(jb)
}
\arguments{
\item{jb}{Flow Job object}
}
\value{
data.table
}
\description{
wrapper for sacct that takes a Flow Job object and returns a data.table containing the time, memory, and state info for each job (SLURM command line function)
}
\author{
Alon Shaiber
}
