% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{sparse_subset}
\alias{sparse_subset}
\title{sparse_subset}
\usage{
sparse_subset(A, B, strict = FALSE, chunksize = 100, quiet = FALSE)
}
\arguments{
\item{A}{k1 x n matrix}

\item{B}{k2 x n matrix}

\item{strict}{logical flag whether to return strict subset (=FALSE)}

\item{chunksize}{integer size of rows to process from each matrix at a single iteration (=100)}

\item{quiet}{logical flag (=FALSE)}
}
\value{
k1 x k2 matrix C whose entries ij = 1 if the set of nonzero components of row i of A is
a (+/- strict) subset of the nonzero components of row j of B
}
\description{
given k1 x n matrix A and k2 x n matrix B
returns k1 x k2 matrix C whose entries ij = 1 if the set of nonzero components of row i of A is
a (+/- strict) subset of the nonzero components of row j of B
}
\author{
Marcin Imielinski
}

