% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{splot}
\alias{splot}
\title{splot}
\usage{
splot(x, y, cex = 0.4, poutlier = 0.01, col = alpha("black", 0.3),
  xlim = quantile(x, na.rm = T, prob = c(poutlier[1], 1 -
  poutlier[length(poutlier)])), ylim = quantile(y, na.rm = T, prob =
  c(poutlier[1], 1 - poutlier[length(poutlier)])), label = NULL,
  cex.label = 1, adj.label = c(1, 0.5), col.label = "black", log = "",
  jiggle = NULL, fit = FALSE, col.fit = "blue", cex.fit = 1,
  square = FALSE, pch = 19, ...)
}
\arguments{
\item{x}{numeric vector of x data}

\item{y}{numeric vector y data}

\item{cex}{character inspection}

\item{poutlier}{numeric value between 0 and 1 specifying quantile threshold of  outliers to remove (=0.01)}

\item{col}{character vector color (=alpha('black', 0.3))}

\item{xlim}{length 2 numeric vector specifying x limits (=quantile(x, na.rm = T, prob = c(poutlier[1], 1-poutlier[length(poutlier)])))}

\item{ylim}{length 2 numeric vector specifying y limits (=quantile(y, na.rm = T, prob = c(poutlier[1], 1-poutlier[length(poutlier)])))}

\item{log}{standard plot log string}

\item{jiggle}{numeric value between 0 and 1 specifying what percentage of plot area to jiggle each point (useful for overplotting) (= NULL)}

\item{fit}{logical flag whether to fit a linear regression line to the data (=FALSE)}

\item{col.fit}{character specifying color of linear regression fit (='blue')}

\item{cex.fit}{character specifying size of text associated with linear regerssion line (=1)}

\item{square}{logical flag whether to make square plot}

\item{pch}{pch}

\item{...}{}
}
\description{
convenient formatted scatter plot with additional features as defaults, useful for fast interactive data inspection / exploration of large
datasets (eg 1000s of points):
- autoamtic setting of solid dots (pch = 19)
- transparent colors for over plotting
- automatic setting of x and y limits parametrized by "p.outlier"
- quick setting of jiggle / jitter on plot
- automatic fitting and plotting of regression line (fit = FALSE)
}
\author{
Marcin Imielinski
}

