% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{ssegment}
\alias{ssegment}
\title{wrapper around DNAcopy to segment a coverage profile}
\usage{
ssegment(cov, field = NULL, log = TRUE, verbose = TRUE, alpha = 1e-05)
}
\arguments{
\item{tcov}{GRanges of binned genome-wide coverage}
}
\value{
GRanges of genomewise segments of piecewise constant coverage
}
\description{
Internal function utilizing DNAcopy to segment a coverage profile
}
\author{
Marcin Imielinski
}
