% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{sstat}
\alias{sstat}
\title{sstat}
\usage{
sstat(full = FALSE, numslots = TRUE, resources = T)
}
\description{
Tabulates Slurm cluster usage (sstat()) or if full = TRUE flag given will dump out
all running jobs in a data.table
}
\author{
Zoran Gajic
}
