% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{stavePDF}
\alias{stavePDF}
\title{stavePDF}
\usage{
stavePDF(object, file, note = NULL, ..., verbose = FALSE)
}
\description{
Stamps and saves PDF file .. i.e. saving datestamped filename and
and soft link to the datestamped file
}
