% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{staveRDS}
\alias{staveRDS}
\title{staveRDS}
\usage{
staveRDS(object, file, note = NULL, ..., verbose = FALSE)
}
\description{
Stamps and saves RDS file .. i.e. saving datestamped filename and
and soft link to the datestamped file
}
