% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{strsplit.fwf}
\alias{strsplit.fwf}
\title{splits a single string according to fixed widths contained in fw (ie each components i of fw denotes the width of field i in string str}
\usage{
strsplit.fwf(str, fw)
}
\description{
splits a single string according to fixed widths contained in fw (ie each components i of fw denotes the width of field i in string str
}

