% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{strsplit2}
\alias{strsplit2}
\title{strsplit2}
\usage{
strsplit2(x, sep1 = ",", sep2 = " ", j = 1)
}
\arguments{
\item{x}{character vector}

\item{sep1}{character specifying first level separator (=',')}

\item{sep2}{character specifying second level separator (=' ')}

\item{j}{integer specifying which subitem to keep (=1)}
}
\value{
vector of values for subitem j
}
\description{
Strsplit when there are two layers of separators (sep1, sep2) and one needs to extract
a collapsed vector of subitem j for all items i.

Takes in a character vector and outputs a list of "separated" items
}
\author{
Marcin Imielinski
}

