% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{sub2ind}
\alias{sub2ind}
\title{sub2ind}
\usage{
sub2ind(dim, r, c, byrow = F)
}
\arguments{
\item{dim}{dimension of matrix to return index for}

\item{r}{integer vector of row index to look up}

\item{c}{length(r) integer vector of column index to look up}

\item{byrow}{whether to calculate indices by row or column (= FALSE)}
}
\value{
length(r) vector of 1D indices into matrix with dim "dim"
}
\description{
MATLAB style sub2ind function in R physical essence.  Provides the one dim matrix index
of row-column locations in matrix

(RIP matlab)
}
\author{
Marcin Imielinski
}

