% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{tabstring}
\alias{tabstring}
\title{tabstring}
\usage{
tabstring(tab, sep = ", ", sep2 = "_", dt = FALSE)
}
\arguments{
\item{tab}{"table" or any named(vector)}

\item{sep}{separator to use between table elements}
}
\value{
character representation of table
}
\description{
string representation of a named vector (ie the result of tab = table(x)
ie name1 (value1), name2 (value2), name3 (value3)
}
\author{
Marcin Imielinski
}

