% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{toggle_grfo}
\alias{.toggle_grfo}
\alias{toggle_grfo}
\title{toggle data.table vs IRanges find overlaps}
\usage{
.toggle_grfo()
}
\description{
toggles global setting of whether to use data.table vs IRanges find overlaps machinery
}
\author{
Marcin Imielinski
}

