% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{ucount}
\alias{ucount}
\title{ucount}
\usage{
ucount(x)
}
\arguments{
\item{x}{vector}
}
\value{
length(x) vector with number of instances of each item in x
}
\description{
returns vector of same length as input with number of counts of each value
in the whole list
}
\author{
Marcin Imielinski
}

