% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{vaggregate}
\alias{vaggregate}
\title{vaggregate}
\usage{
vaggregate(...)
}
\arguments{
\item{...}{arguments to aggregate}
}
\value{
named vector indexed by levels of "by"
}
\description{
same as aggregate except returns named vector
with names as first column of output and values as second

Note: there is no need to ever use aggregate or vaggregate, just switch to data.table
}
\author{
Marcin Imielinski
}

