% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{variants}
\alias{variants}
\title{variants}
\usage{
variants(query, ref, expand.iupac = TRUE, verbose = FALSE)
}
\arguments{
\item{query}{DNAStringSet of query}

\item{ref}{DNASTringSet of ref}

\item{expand.iupac}{logical flag (TRUE) specifying whether to expand iupac for computing SNV  in query and reference}
}
\value{
GRanges in ref coordinates of SNV and indels
}
\description{
Call substitutions and indels from contigs

Calls substitutions from contigs by comparing DNAStringSet ref
against reference sequence ref via RSeqLib::BWA

The caller is IUPAC "ambiguity code aware" meaning that for every instance of ambiguity in the query and reference
it will output the cartesian product of all mismatching variants (iupac = TRUE) otherwise
it will treat those bases literally. 

Note: not recommended to use on reads, only for contigs, i.e. will not scale to millions of reads
}
\author{
Marcin Imielinski
}
