% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{wfplot}
\alias{wfplot}
\title{Quick waterfall plot}
\usage{
wfplot(data, labels = NULL, names.arg = NULL, col = NULL, las = 2,
  cex = 1, leg.pos = NULL, ...)
}
\arguments{
\item{data}{length n numeric vector to be drawn and sorted on y axis}

\item{labels}{length n character vector categorical labels of data}

\item{names.arg}{length n character vector, optional, of individual labels to be drawn verticallyon x axis}

\item{col}{optional named character vector mapping unique category labels to colors}

\item{las}{optional integer vector specifying orientation of labels on barplot}

\item{cex}{numeric value specifying size of names.arg data labels}

\item{leg.pos}{NULL}

\item{...}{additional arguments to barplot}
}
\value{
plot
}
\description{
Quick waterfall plot

data is a numeric vector
labels are text labels of the same length as data
col is either (1) an unamed list of unique colors (2) a named list mapping unique labels to colors
}
\author{
Marcin Imielinski
}
