% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{wij}
\alias{wij}
\title{wij}
\usage{
wij(
  expr,
  filename = "plot.html",
  zoom = NULL,
  cex = 1,
  force = FALSE,
  quiet = FALSE,
  embed = FALSE
)
}
\description{
Evaluates output of htmlwidget generating expression (e.g. via highcharter) and send to filename in predefined WIDGET.DIR
by default plot.html
}
