% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{wijj}
\alias{wijj}
\title{wijj}
\usage{
wijj(x, width = NULL, height = NULL, file = paste0("plotlyJupyterHTML/",
  digest::digest(Sys.time()), ".html"))
}
\description{
Embeds widget in jupyter notebook
}
