% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{write.htab}
\alias{write.htab}
\title{write.htab}
\usage{
write.htab(
  tab,
  file = NULL,
  title = NULL,
  footer = NULL,
  highlight = NULL,
  row.names = TRUE,
  col.names = TRUE,
  high.color = "yellow",
  row.colors = c("lightgray", "white"),
  header.colors = c("#4A4A4A", "white"),
  data.size = 15,
  dt = TRUE,
  force = FALSE,
  embed = FALSE,
  title.size = 15,
  footer.size = 20,
  header.size = round(1.1 * data.size)
)
}
\arguments{
\item{tab}{data.frame, data.table, or GRanges}

\item{file}{optional output .html file (by default ~/public_html/htab.html but directory can be set using env variable HTAB.PATH)}

\item{title}{title to the page (=NULL)}

\item{footer}{footer to add to page (=NULL)}

\item{highlight}{optional integer vector specifying what rows to highlight}

\item{row.names}{logical flag whether to include row labels (=TRUE)}

\item{col.names}{logical flag whether to include col labels (=TRUE)}

\item{high.color}{chaoracter highlight color (= 'yellow')}

\item{row.colors}{length 2 character vector to shade data rows (= c('lightgray', 'white'))}

\item{header.colors}{length 2 character vector specifygin background and text for header row (= c('#4A4A4A', 'white'))}

\item{data.size}{integer font size in px for data, title, and footer (= 15)}

\item{title.size}{integer font size in px for title (= 15)}

\item{footer.size}{integer font size in px for footer (= 15)}

\item{header.size}{integer font size in px for header (= 15)}
}
\description{
writes data frame (or anything castable to data frame) to pretty HTML formatted table to a static location.
Useful for quick table inspection via Chrome or other web browser.

Very similar syntax to write.tab
}
\author{
Marcin Imielinski
}
