% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skitools.R
\name{write.tab}
\alias{write.tab}
\title{write.tab
writes tab delimited no quotes without row names table (passes remaining arguments to write.table)

equivalent to write.table(sep = TAB.DELIM, quote = F, row.names = F)}
\usage{
write.tab(x, ..., sep = "\\t", quote = F, row.names = F)
}
\arguments{
\item{x}{data.frame to dump}

\item{...}{additional arguments to write.table}
}
\description{
write.tab
writes tab delimited no quotes without row names table (passes remaining arguments to write.table)

equivalent to write.table(sep = TAB.DELIM, quote = F, row.names = F)
}

