% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{write_ggjs}
\alias{write_ggjs}
\title{dumps ggjs formatted csv file from GRanges}
\usage{
write_ggjs(gr, filename, field = "score")
}
\arguments{
\item{gr}{GRanges input}

\item{filename}{filename to dump csv to}

\item{field}{field name to dump as (numeric) data ["score"]}
}
\description{

}
\author{
Marcin Imielinski
}
