% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{alpha}
\alias{alpha}
\title{alphoa}
\usage{
alpha(col, alpha)
}
\arguments{
\item{col}{RGB color}

\item{alpha}{value of alpha}
}
\description{
Give transparency value to colors

Takes provided colors and gives them the specified alpha (ie transparency) value
}
\author{
Marcin Imielinski
}

