% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{dplot}
\alias{dplot}
\title{dplot}
\usage{
dplot(
  y,
  group,
  ylab = "",
  xlab = "",
  log = F,
  dotsize = NULL,
  binwidth = 0.02,
  title = NULL,
  ylim = NULL,
  text.size = NULL
)
}
\arguments{
\item{y}{numeric vector of data}

\item{group}{length(y) vector of catageories}

\item{ylab}{y axis label (='')}

\item{xlab}{x axis label (='')}

\item{log}{logical flag whether to plot y axis in log format (=FALSE)}

\item{dotsize}{integer dot size to plot with, as function of 0.02 category width plot real estate (= NULL)}

\item{binwidth}{numeric binwidth of histogram in units of data quantiles (= NULL)}

\item{title}{character title of plot (='')}

\item{ylim}{y limits of plot (= NULL)}

\item{text.size}{text size of legend (= NULL)}
}
\description{
Plots dotplot of grouped data
}
\author{
Marcin Imielinski
}
