% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{grok_vcf}
\alias{grok_vcf}
\title{grok_vcf}
\usage{
grok_vcf(x, label = NA, keep.modifier = TRUE, long = FALSE,
  oneliner = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{GRanges input}
}
\description{
Does additional processing of annotated vcf output and produces
a more readable granges output.
}
