% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{hets}
\alias{hets}
\title{hets}
\usage{
hets(tum.bam, norm.bam = NULL, out.file,
  vcf.file = "/cga/meyerson/home/marcin/DB/dbSNP/hapmap_3.3.b37.vcf",
  chunk.size1 = 1000, chunk.size2 = 100, mc.cores = 1, verbose = T,
  na.rm = TRUE, filt.norm = T)
}
\description{
generates allele fraction at all possible hets at sites specified by vcf (eg hapmap) input
for tumor and normal
}
