% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{ra.overlaps}
\alias{ra.overlaps}
\alias{ra.setdiff}
\title{identifies events that are in ra1 that do not exist in ra2.
Aside from ra1 and ra2, all arguments are sent to ra.overlaps}
\usage{
ra.setdiff(ra1, ra2, ...)
}
\description{
identifies events that are in ra1 that do not exist in ra2.
Aside from ra1 and ra2, all arguments are sent to ra.overlaps
}

