% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skitools.R
\name{vplot}
\alias{vplot}
\title{vplot}
\usage{
vplot(y, group = "x", facet1 = NULL, facet2 = NULL, transpose = FALSE,
  mapping = NULL, stat = "ydensity", position = "dodge", trim = TRUE,
  scale = "area", log = FALSE, count = TRUE, xlab = NULL, ylim = NULL,
  ylab = NULL, minsup = NA, scatter = FALSE, text = NULL,
  cex.scatter = 1, col.scatter = NULL, alpha = 0.3, title = NULL,
  legend.ncol = NULL, legend.nrow = NULL, vfilter = TRUE, vplot = TRUE,
  dot = FALSE, stackratio = 1, binwidth = 0.1, plotly = FALSE,
  print = TRUE)
}
\arguments{
\item{y}{numeric data vector}

\item{group}{length(y) vector of categories}

\item{facet1}{optional length(y) vector of row categories to facet on (=NULL)}

\item{facet2}{optional length(y) vector of column categories to facet on (=NULL)}

\item{transpose}{logical vector whether flip row / column orientation of facets (=FALSE)}

\item{mapping}{mapping of groups to colors (=NULL)}

\item{scale}{scale parameter to geom_vplot (=area)}

\item{log}{logical flag whether to log transform (=FALSE)}

\item{count}{logical flag whether to include counts in ylabels (=TRUE)}

\item{xlab}{character xlabel (=NULL)}

\item{ylab}{character ylabel (=NULL)}

\item{minsup}{minimum support to include in a group (=NA)}

\item{scatter}{logical flag whether to include scatter of points (=FALSE)}

\item{alpha}{numeric vector between 0 and 1 to specify alpha transparency of points if scatter is TRUE (0.3)}

\item{title}{character specifying plot title (=NULL)}
}
\description{
Quick violin plot
}
\author{
Marcin Imielinski
}

